### 2023-04-03 #5332 @NickleDave
Just commenting to make sure it doesn't get missed that this was previously reviewed by pyOpenSci: https://github.com/pyOpenSci/software-submission/issues/68

For a previous (recent) JOSS submission following a pyOpenSci review see https://github.com/openjournals/joss-reviews/issues/5289
