### 2023-04-05 #5338 @editorialbot
[REVIEW]: crowsetta: A Python tool to work with any format for annotating animal vocalizations and bioacoustics data
**Submitting author:** @NickleDave (<a href="http://orcid.org/0000-0002-4261-4719">David Nicholson</a>)
**Repository:** https://github.com/vocalpy/crowsetta
**Branch with paper.md** (empty if default branch): add-joss-paper
**Version:** 5.0.0
