### 2023-04-06 #5338 @NickleDave
Hi @oliviaguest I verified the suggested DOIs under "MISSING DOIs" were correct and added them to paper.bib.
Seems like the PDF regenerates ok using the GitHub Actions workflow on my repo 🤞

Just let me know if there's any other changes you need me to make.
